# openapi.yaml
openapi: 3.1.0
info:
  title: Gatekeeper
  description: >
    API gateway on Cloudflare Workers with AWS IAM-style policy authorization.
    Fronts Cloudflare's cache purge API and S3-compatible R2 storage with
    per-key/credential policies, token-bucket rate limiting, request collapsing,
    and D1-backed analytics. The S3 proxy provides fine-grained object-level
    access control that R2's native tokens cannot — key-prefix restrictions,
    per-operation permissions, cross-bucket mixed policies.
  version: 3.0.0
  contact:
    name: Erfi Anugrah
servers:
  - url: https://gate.erfi.io
    description: Production

# ─── Security schemes ──────────────────────────────────────────────────────
security: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: >
        Gateway API key (prefix `gw_`). Issued via `POST /admin/keys`.
        Each key has an attached IAM policy that controls which actions,
        resources, and conditions it can access.

    AdminKeyAuth:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: >
        Shared admin secret for CLI and automation. Falls back when
        Cloudflare Access is not configured.

    CloudflareAccess:
      type: apiKey
      in: header
      name: Cf-Access-Jwt-Assertion
      description: >
        Cloudflare Access JWT. Injected automatically by Access for
        browser-based dashboard users. Also accepted via `CF_Authorization`
        cookie.

    S3SigV4Auth:
      type: apiKey
      in: header
      name: Authorization
      description: >
        AWS Signature Version 4 authentication. S3 clients (rclone, boto3,
        aws-cli) send this automatically. Credentials are issued via
        `POST /admin/s3/credentials` (access key prefix `GK`). Also
        supports presigned URL authentication via query parameters.

  # ─── Reusable schemas ──────────────────────────────────────────────────
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [success, errors]
      properties:
        success:
          type: boolean
          const: false
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ApiError'
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
        message:
          type: string

    # --- Bulk operation results ---
    BulkItemResult:
      type: object
      required: [id, status]
      properties:
        id:
          type: string
        status:
          type: string
          enum: [revoked, deleted, already_revoked, not_found]
    BulkResult:
      type: object
      required: [processed, results]
      properties:
        processed:
          type: integer
          description: Number of items processed.
        results:
          type: array
          items:
            $ref: '#/components/schemas/BulkItemResult'
    BulkDryRunItem:
      type: object
      required: [id, current_status, would_become]
      properties:
        id:
          type: string
        current_status:
          type: string
          enum: [active, revoked, expired, not_found]
        would_become:
          type: string
    BulkDryRunResult:
      type: object
      required: [dry_run, would_process, items]
      properties:
        dry_run:
          type: boolean
          const: true
        would_process:
          type: integer
          description: Number of items that would be processed.
        items:
          type: array
          items:
            $ref: '#/components/schemas/BulkDryRunItem'

    # --- Policy document ---
    PolicyDocument:
      type: object
      required: [version, statements]
      properties:
        version:
          type: string
          const: '2025-01-01'
        statements:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/Statement'
    Statement:
      type: object
      required: [effect, actions, resources]
      properties:
        effect:
          type: string
          enum: [allow, deny]
          description: >
            Whether this statement allows or denies matching requests.
            Deny-first: if any deny statement matches, the request is denied
            regardless of allow statements.
        actions:
          type: array
          minItems: 1
          items:
            type: string
          description: >
            Actions this statement permits or denies. Wildcard suffix supported.
            Purge actions: `purge:url`, `purge:host`, `purge:tag`,
            `purge:prefix`, `purge:everything`, `purge:*`.
          examples:
            - ['purge:url', 'purge:host']
            - ['purge:*']
        resources:
          type: array
          minItems: 1
          items:
            type: string
          description: >
            Resources this statement applies to. Format: `zone:<id>`.
            Wildcard: `zone:*` or `*`.
          examples:
            - ['zone:aaaa1111bbbb2222cccc3333dddd4444']
            - ['zone:*']
        conditions:
          type: array
          items:
            $ref: '#/components/schemas/Condition'
          description: >
            Optional conditions — all must match (AND).
            Omit or empty array for no conditions.
    Condition:
      oneOf:
        - $ref: '#/components/schemas/LeafCondition'
        - $ref: '#/components/schemas/AnyCondition'
        - $ref: '#/components/schemas/AllCondition'
        - $ref: '#/components/schemas/NotCondition'
...
    AnyCondition:
      type: object
      required: [any]
      properties:
        any:
          type: array
          items:
            $ref: '#/components/schemas/Condition'
          description: OR — any child must match.
    AllCondition:
      type: object
      required: [all]
      properties:
        all:
          type: array
          items:
            $ref: '#/components/schemas/Condition'
          description: AND — all children must match (explicit).
...
    ApiKey:
      type: object
      required: [id, name, zone_id, created_at, revoked, policy]
      properties:
        id:
          type: string
          examples: ['gw_a1b2c3d4e5f6']
        name:
          type: string
        zone_id:
          type: string
          pattern: '^[a-f0-9]{32}$'
        created_at:
          type: integer
          description: Unix timestamp (ms).
        expires_at:
          type: ['integer', 'null']
          description: Unix timestamp (ms), or null if no expiration.
        revoked:
          type: integer
          enum: [0, 1]
          description: 0 = active, 1 = revoked.
        policy:
          type: string
          description: JSON-serialized PolicyDocument.
        created_by:
          type: ['string', 'null']
          description: Email of the creator (from Access JWT), or null.
        bulk_rate:
          type: ['integer', 'null']
          description: Per-key bulk rate limit (req/sec). Null = account default.
        bulk_bucket:
          type: ['integer', 'null']
        single_rate:
          type: ['integer', 'null']
          description: Per-key single-file rate limit (URLs/sec). Null = account default.
        single_bucket:
          type: ['integer', 'null']

    # --- Create key request ---
    CreateKeyRequest:
      type: object
      required: [name, zone_id, policy]
      properties:
        name:
          type: string
        zone_id:
          type: string
          pattern: '^[a-f0-9]{32}$'
        policy:
          $ref: '#/components/schemas/PolicyDocument'
        expires_in_days:
          type: integer
          minimum: 1
        created_by:
          type: string
          description: >
            Auto-populated from Access JWT email when authenticated
            via Cloudflare Access. Can be set manually when using admin key.
        rate_limit:
          type: object
          properties:
            bulk_rate:
              type: integer
              description: Must not exceed account default.
            bulk_bucket:
              type: integer
            single_rate:
              type: integer
            single_bucket:
              type: integer

    # --- Purge body ---
...
    GatewayConfig:
      type: object
      required:
        - bulk_rate
        - bulk_bucket_size
        - bulk_max_ops
        - single_rate
        - single_bucket_size
        - single_max_ops
        - key_cache_ttl_ms
        - retention_days
      properties:
        bulk_rate:
          type: number
          description: Bulk purge rate limit (req/sec).
        bulk_bucket_size:
          type: number
          description: Bulk purge token bucket capacity.
        bulk_max_ops:
          type: number
          description: Max operations per bulk purge request.
        single_rate:
          type: number
          description: Single-file purge rate limit (URLs/sec).
        single_bucket_size:
          type: number
          description: Single-file purge token bucket capacity.
        single_max_ops:
          type: number
          description: Max URLs per single-file purge request.
        key_cache_ttl_ms:
          type: number
          description: API key cache TTL in milliseconds.
        retention_days:
          type: number
          description: Analytics event retention in days.
    ConfigOverride:
      type: object
      required: [key, value, updated_at]
      properties:
        key:
          type: string
        value:
          type: string
        updated_at:
          type: integer
          description: Unix timestamp (ms).
        updated_by:
          type: ['string', 'null']
    ConfigResponse:
      type: object
      required: [config, overrides, defaults]
      properties:
        config:
          $ref: '#/components/schemas/GatewayConfig'
        overrides:
          type: array
          items:
            $ref: '#/components/schemas/ConfigOverride'
        defaults:
          type: object
          additionalProperties:
            type: number

  # ─── Reusable parameters ─────────────────────────────────────────────────
  parameters:
    ZoneIdQuery:
      name: zone_id
      in: query
      required: true
      schema:
        type: string
        pattern: '^[a-f0-9]{32}$'
      description: Cloudflare zone ID.
...
    KeyIdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: API key ID.
...
paths:
...
  /admin/keys:
    post:
      operationId: createKey
      summary: Create API key
      description: >
        Create a new gateway API key with an attached IAM policy document.
        The returned key `id` (prefix `gw_`) is the Bearer token used
        for purge requests.
      tags: [Keys]
      security:
        - CloudflareAccess: []
        - AdminKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateKeyRequest'
            examples:
              minimal:
                summary: Full purge access on one zone
                value:
                  name: ci-deploy
                  zone_id: aaaa1111bbbb2222cccc3333dddd4444
                  policy:
                    version: '2025-01-01'
                    statements:
                      - effect: allow
                        actions: ['purge:*']
                        resources: ['zone:aaaa1111bbbb2222cccc3333dddd4444']
              scoped:
                summary: URL purge restricted to specific hosts
                value:
                  name: cdn-purger
                  zone_id: aaaa1111bbbb2222cccc3333dddd4444
                  policy:
                    version: '2025-01-01'
                    statements:
                      - effect: allow
                        actions: ['purge:url']
                        resources: ['zone:aaaa1111bbbb2222cccc3333dddd4444']
                        conditions:
                          - field: host
                            operator: ends_with
                            value: '.example.com'
                  expires_in_days: 90
              deny-off-hours:
                summary: Allow purge but deny purge_everything outside business hours
                value:
                  name: safe-purger
                  zone_id: aaaa1111bbbb2222cccc3333dddd4444
                  policy:
                    version: '2025-01-01'
                    statements:
                      - effect: allow
                        actions: ['purge:*']
                        resources: ['zone:aaaa1111bbbb2222cccc3333dddd4444']
                      - effect: deny
                        actions: ['purge:everything']
                        resources: ['zone:*']
                        conditions:
                          - field: time.hour
                            operator: gte
                            value: '22'
      responses:
        '200':
          description: Key created successfully.
          content:
            application/json:
              schema:
                type: object
                required: [success, result]
                properties:
                  success:
                    type: boolean
                    const: true
                  result:
                    type: object
                    required: [key]
                    properties:
                      key:
                        $ref: '#/components/schemas/ApiKey'
                        description: >
                          The created API key. The `id` field (prefix `gw_`)
                          is used as the Bearer token for purge requests.
        '400':
          description: Validation error (missing fields, invalid policy, rate limits exceed account defaults).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

    get:
      operationId: listKeys
      summary: List API keys
      tags: [Keys]
      security:
        - CloudflareAccess: []
        - AdminKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/ZoneIdQuery'
        - name: status
          in: query
          schema:
            type: string
            enum: [active, revoked]
          description: Filter by key status.
      responses:
        '200':
          description: List of API keys for the zone.
          content:
            application/json:
              schema:
                type: object
                required: [success, result]
                properties:
                  success:
                    type: boolean
                    const: true
                  result:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApiKey'
        '400':
          description: Missing zone_id query parameter.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /admin/keys/{id}:
    get:
      operationId: getKey
      summary: Get API key details
      description: Returns the key metadata including the serialized policy.
      tags: [Keys]
      security:
        - CloudflareAccess: []
        - AdminKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/KeyIdPath'
        - $ref: '#/components/parameters/ZoneIdQuery'
      responses:
        '200':
          description: Key details.
          content:
            application/json:
              schema:
                type: object
                required: [success, result]
                properties:
                  success:
                    type: boolean
                    const: true
                  result:
                    type: object
                    required: [key]
                    properties:
                      key:
                        $ref: '#/components/schemas/ApiKey'
        '400':
          description: Missing zone_id query parameter.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Key not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

    delete:
      operationId: revokeOrDeleteKey
      summary: Revoke or permanently delete an API key
      description: >
        Without `permanent`, soft-revokes the key (sets `revoked = 1`).
        With `permanent=true`, permanently removes the key from the database.
        Hard-delete works on keys in any state (active or already revoked).
        Analytics events referencing the key are preserved.
      tags: [Keys]
      security:
        - CloudflareAccess: []
        - AdminKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/KeyIdPath'
        - $ref: '#/components/parameters/ZoneIdQuery'
        - name: permanent
          in: query
          required: false
          schema:
            type: boolean
          description: >
            When `true`, permanently deletes the key row instead of soft-revoking.
      responses:
        '200':
          description: Key revoked or deleted.
          content:
            application/json:
              schema:
                type: object
                required: [success, result]
                properties:
                  success:
                    type: boolean
                    const: true
                  result:
                    type: object
                    properties:
                      revoked:
                        type: boolean
                        description: Present when soft-revoking.
                      deleted:
                        type: boolean
                        description: Present when permanently deleting.
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Key not found (or already revoked, when soft-revoking).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  # ─── Admin: Bulk Key Operations ───────────────────────────────────────────
  /admin/keys/bulk-revoke:
    post:
      operationId: bulkRevokeKeys
      summary: Bulk soft-revoke multiple API keys
      description: >
        Soft-revoke multiple keys in one request. Requires `confirm_count` to match the
        number of IDs as a fat-finger guard. Set `dry_run: true` to preview without executing.
      tags: [Keys]
      security:
        - CloudflareAccess: []
        - AdminKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids, confirm_count]
              properties:
                ids:
                  type: array
                  items:
                    type: string
                  minItems: 1
                  maxItems: 100
                  description: Array of key IDs to revoke.
                confirm_count:
                  type: integer
                  description: Must exactly match the length of `ids`.
                dry_run:
                  type: boolean
                  default: false
                  description: Preview the operation without executing.
      responses:
        '200':
          description: Bulk revoke result or dry-run preview.
          content:
            application/json:
              schema:
                type: object
                required: [success, result]
                properties:
                  success:
                    type: boolean
                    const: true
                  result:
                    oneOf:
                      - $ref: '#/components/schemas/BulkResult'
                      - $ref: '#/components/schemas/BulkDryRunResult'
        '400':
          description: Validation error (empty array, confirm_count mismatch, over 100 items).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /admin/keys/bulk-delete:
    post:
      operationId: bulkDeleteKeys
      summary: Bulk permanently delete multiple API keys
      description: >
        Permanently remove multiple keys from the database. Works on keys in any state.
        Requires `confirm_count` as a fat-finger guard. Set `dry_run: true` to preview.
        Analytics events referencing deleted keys are preserved.
      tags: [Keys]
      security:
        - CloudflareAccess: []
        - AdminKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids, confirm_count]
              properties:
                ids:
                  type: array
                  items:
                    type: string
                  minItems: 1
                  maxItems: 100
                  description: Array of key IDs to delete.
                confirm_count:
                  type: integer
                  description: Must exactly match the length of `ids`.
                dry_run:
                  type: boolean
                  default: false
                  description: Preview the operation without executing.
      responses:
        '200':
          description: Bulk delete result or dry-run preview.
          content:
            application/json:
              schema:
                type: object
                required: [success, result]
                properties:
                  success:
                    type: boolean
                    const: true
                  result:
                    oneOf:
                      - $ref: '#/components/schemas/BulkResult'
                      - $ref: '#/components/schemas/BulkDryRunResult'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  # ─── Admin: Analytics ─────────────────────────────────────────────────────
...
  /admin/config:
    get:
      operationId: getConfig
      summary: Get gateway configuration
      description: >
        Returns the resolved configuration, active overrides, and defaults.
      tags: [Config]
      security:
        - CloudflareAccess: []
        - AdminKeyAuth: []
      responses:
        '200':
          description: Current configuration.
          content:
            application/json:
              schema:
                type: object
                required: [success, result]
                properties:
                  success:
                    type: boolean
                    const: true
                  result:
                    $ref: '#/components/schemas/ConfigResponse'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

    put:
      operationId: setConfig
      summary: Update gateway configuration
      description: >
        Set one or more config values. Values must be positive finite numbers.
        Only known config keys are accepted.
      tags: [Config]
      security:
        - CloudflareAccess: []
        - AdminKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: number
              description: >
                Key-value pairs of config keys to update.
              examples:
                - bulk_rate: 50
                  retention_days: 30
      responses:
        '200':
          description: Configuration updated.
          content:
            application/json:
              schema:
                type: object
                required: [success, result]
                properties:
                  success:
                    type: boolean
                    const: true
                  result:
                    type: object
                    required: [config]
                    properties:
                      config:
                        $ref: '#/components/schemas/GatewayConfig'
        '400':
          description: Validation error (unknown key, invalid value).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...